+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this package will be documented in this file.
 
+## v0.2.0
+
+* switch to manager driver pattern
+  * `\Elbgoods\TrashmailRule\TrashmailManager`
+  * `\Elbgoods\TrashmailRule\Providers`
+* add service class `\Elbgoods\TrashmailRule\Trashmail`
+* add facade `\Elbgoods\TrashmailRule\Facades\Trashmail`
+* try-catch all errors in providers and skip this provider
+
 ## v0.1.0
 
 * initial release